backend/requirements.txt
# Production Dependencies
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0  # ADDED THIS
prometheus-client==0.19.0
blinker==1.7.0
python-dateutil==2.8.2
pyjwt==2.8.0
